examples/video/requirements.txt
opencv-python
Pillow
torch
torchvision
imageai
moviepy
#flexecutor
